pyproject.toml
[tool.poetry]
name = "verbiste-tools"
version = "0.1.0"
description = "Support scripts for verbiste.el"
authors = ["Aidan Pace <apace@defrecord.com>"]
readme = "README.md"
license = "GPL-2.0-or-later"
packages = [{include = "verbiste_tools"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
lxml = "^4.9.3"
numpy = "^1.26.0"
scikit-learn = "^1.3.2"
matplotlib = "^3.8.0"
pandas = "^2.1.1"
spacy = "^3.8.5"
google-genai = "1.7.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
black = "^23.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
xml2json = "verbiste_tools.xml2json:main"
cluster_verbs = "verbiste_tools.cluster_verbs:main"

[tool.black]
line-length = 88
target-version = ["py39"]
